# artifacts/critiq/src/types/api.ts
// API response types matching the backend

export type ApiCategory = {
  id: number;
  slug: string;
  name: string;
  icon: string;
  sortOrder: number;
};

export type ApiCriterion = {
  id: number;
  categoryId: number;
  name: string;
  description: string | null;
  status: string;
  sortOrder: number;
};

export type ApiRatingEntry = {
  criterionId: number;
  score: number;
  count: number;
};

export type ApiProduct = {
  id: number;
  categoryId: number;
  name: string;
  brand: string;
  modelNumber: string;
  janCode: string | null;
  price: number;
  description: string | null;
  status: string;
  reviewCount: number;
  images: string[];
  ratings: ApiRatingEntry[];
};

export type AdminProductItem = {
  id: number;
  categoryId: number;
  name: string;
  brand: string;
  modelNumber: string;
  janCode: string | null;
  price: number;
  description: string | null;
  status: string;
  reviewCount: number;
  images: string[];
  createdAt: string;
};

export type CriterionSuggestionItem = {
  id: number;
  categoryId: number;
  name: string;
  description: string;
  reason: string | null;
  status: string;
  adminNotes: string | null;
  createdAt: string;
};

export type ProductSuggestionItem = {
  id: number;
  categoryId: number;
  brand: string;
  name: string;
  modelNumber: string;
  janCode: string | null;
  price: number | null;
  description: string | null;
  images: string[];
  status: string;
  adminNotes: string | null;
  createdAt: string;
};