# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface CategoryItem {
  id: number;
  slug: string;
  name: string;
  icon: string;
  sortOrder: number;
}

export interface CategoryInput {
  slug: string;
  name: string;
  icon: string;
  sortOrder?: number;
}

export interface CriterionItem {
  id: number;
  categoryId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  sortOrder: number;
}

export interface CriterionInput {
  categoryId: number;
  name: string;
  description?: string;
  status?: string;
  sortOrder?: number;
}

export interface ProductRatingEntry {
  criterionId: number;
  score: number;
  count: number;
}

export interface ProductWithRatings {
  id: number;
  categoryId: number;
  name: string;
  brand: string;
  modelNumber: string;
  /** @nullable */
  janCode?: string | null;
  price: number;
  /** @nullable */
  description?: string | null;
  status: string;
  reviewCount: number;
  images: string[];
  ratings: ProductRatingEntry[];
}

export interface AdminProductItem {
  id: number;
  categoryId: number;
  name: string;
  brand: string;
  modelNumber: string;
  /** @nullable */
  janCode?: string | null;
  price: number;
  /** @nullable */
  description?: string | null;
  status: string;
  reviewCount: number;
  images: string[];
  createdAt?: string;
}

export interface AdminProductInput {
  categoryId: number;
  name: string;
  brand: string;
  modelNumber?: string;
  janCode?: string;
  price?: number;
  description?: string;
  status?: string;
  images?: string[];
}

export type RatingInputScores = {[key: string]: number};

export interface RatingInput {
  scores: RatingInputScores;
}

export interface StatusUpdate {
  status: string;
  adminNotes?: string;
}

export interface CriterionSuggestionInput {
  categoryId: number;
  name: string;
  description: string;
  reason?: string;
}

export interface CriterionSuggestionItem {
  id: number;
  categoryId: number;
  name: string;
  description: string;
  /** @nullable */
  reason?: string | null;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
}

export interface ProductSuggestionInput {
  categoryId: number;
  brand: string;
  name: string;
  modelNumber?: string;
  janCode?: string;
  price?: number;
  description?: string;
  images?: string[];
}

export interface ProductSuggestionItem {
  id: number;
  categoryId: number;
  brand: string;
  name: string;
  modelNumber?: string;
  /** @nullable */
  janCode?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  description?: string | null;
  images?: string[];
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
}

export interface JanLookupResult {
  found: boolean;
  name?: string;
  brand?: string;
  description?: string;
  images?: string[];
  /** @nullable */
  lowestPrice?: number | null;
  /** @nullable */
  highestPrice?: number | null;
}

export interface AdminLoginInput {
  password: string;
}

export interface SuggestionReview {
  status: string;
  adminNotes?: string;
}

export type AdminListCriteriaParams = {
categoryId?: number;
};

export type AdminListProductsParams = {
status?: string;
categoryId?: number;
};

export type AdminListCriterionSuggestionsParams = {
status?: string;
};

export type AdminListProductSuggestionsParams = {
status?: string;
};