# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRITIQ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin auth status
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all categories (admin)
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryItem"
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryItem"
  /admin/categories/{categoryId}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/criteria:
    get:
      operationId: adminListCriteria
      tags: [admin]
      summary: List all criteria (admin, all statuses)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CriterionItem"
    post:
      operationId: adminCreateCriterion
      tags: [admin]
      summary: Create a criterion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriterionItem"
  /admin/criteria/{criterionId}:
    put:
      operationId: adminUpdateCriterion
      tags: [admin]
      summary: Update a criterion
      parameters:
        - name: criterionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriterionItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteCriterion
      tags: [admin]
      summary: Delete a criterion
      parameters:
        - name: criterionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products (admin, all statuses)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProductItem"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a product directly (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductItem"
  /admin/products/{productId}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{productId}/status:
    patch:
      operationId: adminUpdateProductStatus
      tags: [admin]
      summary: Approve or reject a pending product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CategoryItem:
      type: object
      required: [id, slug, name, icon, sortOrder]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    CategoryInput:
      type: object
      required: [slug, name, icon]
      properties:
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    CriterionItem:
      type: object
      required: [id, categoryId, name, status, sortOrder]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        sortOrder:
          type: integer
    CriterionInput:
      type: object
      required: [categoryId, name]
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        sortOrder:
          type: integer
...
    AdminProductItem:
      type: object
      required: [id, categoryId, name, brand, modelNumber, price, status, reviewCount, images]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        brand:
          type: string
        modelNumber:
          type: string
        janCode:
          type: ["string", "null"]
        price:
          type: integer
        description:
          type: ["string", "null"]
        status:
          type: string
        reviewCount:
          type: integer
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
    AdminProductInput:
      type: object
      required: [categoryId, name, brand]
      properties:
        categoryId:
          type: integer
        name:
          type: string
        brand:
          type: string
        modelNumber:
          type: string
        janCode:
          type: string
        price:
          type: integer
        description:
          type: string
        status:
          type: string
        images:
          type: array
          items:
            type: string
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        adminNotes:
          type: string
...
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
...