# lib/api-zod/src/generated/types/adminProductInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminProductInput {
  categoryId: number;
  name: string;
  brand: string;
  modelNumber?: string;
  janCode?: string;
  price?: number;
  description?: string;
  status?: string;
  images?: string[];
}