# lib/api-zod/src/generated/types/adminProductItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminProductItem {
  id: number;
  categoryId: number;
  name: string;
  brand: string;
  modelNumber: string;
  /** @nullable */
  janCode?: string | null;
  price: number;
  /** @nullable */
  description?: string | null;
  status: string;
  reviewCount: number;
  images: string[];
  createdAt?: string;
}