# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  slug: string;
  name: string;
  icon: string;
  sortOrder?: number;
}