# lib/api-zod/src/generated/types/criterionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface CriterionInput {
  categoryId: number;
  name: string;
  description?: string;
  status?: string;
  sortOrder?: number;
}