# lib/api-zod/src/generated/types/criterionItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CRITIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface CriterionItem {
  id: number;
  categoryId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: string;
  sortOrder: number;
}